pyproject.toml
[project]
name = "bpg"
version = "0.1.0"
description = "Business Process Graph — declarative, infrastructure-as-code system for business workflows"
authors = [
    { name = "ginstrom", email = "r-ginstrom@mercari.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
bpg = "bpg.cli:app"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
